+++ src/Requests/Products/Handlers/FindProductsRequestHandler.php
@@ -3,6 +3,8 @@
 use Ordercloud\Entities\Products\ProductCollection;
 use Ordercloud\Requests\Handlers\AbstractPutRequestHandler;
 use Ordercloud\Requests\Handlers\FormatBooleanTrait;
+use Ordercloud\Requests\Products\Criteria\ProductCriteria;
+use Ordercloud\Requests\Products\Criteria\ProductTagCriteria;
 use Ordercloud\Requests\Products\FindProductsRequest;
 use Ordercloud\Support\Reflection\EntityReflector;
 
@@ -34,8 +36,8 @@ class FindProductsRequestHandler extends AbstractPutRequestHandler
                  'organisations' => $criteria->getOrganisations(),
              ])
             ->setQueryParameters([
-                "page"     => $criteria->getPage(),
-                "pagesize" => $criteria->getPageSize()
+                'page'     => $criteria->getPage(),
+                'pagesize' => $criteria->getPageSize()
             ])
              ->setEntityArrayClass('Ordercloud\Entities\Products\Product');